third_party/blink/web_tests/virtual/prefetch/external/wpt/speculation-rules/prefetch/tentative/service-worker/basic.sub.https_origin=same-site_sw=race-fetch-handler-modify-url-expected.txt

This is a testharness.js-based test.
[FAIL] Prefetch with ServiceWorker (race-fetch-handler-modify-url)
  assert_equals: a prefetch request without Sec-Purpose should be sent to the server. expected 1 but got 2
Harness: the test ran to completion.
